fullscreen mode as own screen, still some bugs
