+++ app/app/main/explore/page/[pageId]/fullscreen/_layout.tsx
@@ -1,0 +1,4 @@
+import { Fragment } from "react";
+import { withFullscreenLayout } from "../_layout-base";
+
+export default withFullscreenLayout(Fragment, true);
