+++ app/app/main/explore/page/[pageId]/fullscreen/index.tsx
@@ -1,0 +1,2 @@
+import { generatePage } from "../_main-base";
+export default generatePage(true);
