+++ app/app/main/explore/page/[pageId]/map/index.tsx
@@ -1,0 +1,3 @@
+import { generatePage } from "../_main-base";
+
+export default generatePage(false);
