+++ app/app/main/explore/page/[pageId]/map/section/[title].tsx
@@ -1,0 +1,3 @@
+import { generateSection } from "../../_section-base";
+
+export default generateSection(false);
