+++ app/components/place-card.tsx
@@ -12,11 +12,11 @@ export default function PlaceCard({ item }: { item: Place }) {
   const router = useRouter();
 
   function openPlace() {
-    router.navigate(`/main/explore/page/${item.id}`);
+    router.navigate(`/main/explore/page/${item.id}/map`);
   }
 
   return (
-    <Pressable onPress={openPlace} className="mb-1">
+    <Pressable onPress={openPlace} className="mb -1">
       <Card variant="elevated" size="sm" className="rounded-lg">
         <Heading size="lg">{item.title}</Heading>
       </Card>
