+++ app/hooks/use-move-to.ts
@@ -4,7 +4,6 @@ export const CameraRefContext = createContext<
   (lng: number, lat: number, zoom: number) => void
 >(() => {
   // noop
-  console.log("noop, inside defaultValue");
 });
 
 export default function useMoveTo() {
